% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_of_hydro_year.R
\name{day_of_hydro_year}
\alias{day_of_hydro_year}
\title{Day of hydrologic year}
\usage{
day_of_hydro_year(hDates = NULL)
}
\arguments{
\item{hDates}{date or date vector
day_of_hydro_year()}
}
\description{
This function reads a date or vector of dates and returns the day of the
hydrologic year (begining Oct 01)
}
