% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_year.R
\name{hydro_year}
\alias{hydro_year}
\title{Day of hydrologic year}
\usage{
hydro_year(hDates = NULL)
}
\arguments{
\item{hDates}{date or date vector}
}
\description{
This function reads a date or vector of dates and returns the hydrologic year
(begining Oct 01)
}
